"cambio en update"

+++ Models/Usuarios.php
@@ -63,13 +63,13 @@ class Usuarios{
     //Registro de usuario
     public function createUser($data, $membresia = null)
     {
-        $query = 'INSERT INTO usuarios (dni_usuario,nombre_usuario,nombreCompleto,num_telefono, email, password) VALUES (?,?,?)';
+        $query = 'INSERT INTO usuarios (dni_usuario,nombre_usuario,nombreCompleto,num_telefono, email, password) VALUES (?,?,?,?,?,?)';
         
         try {
             $hashed_pass = password_hash($data['password'], PASSWORD_DEFAULT);
 
             $stmt = $this->con->prepare($query);
-            $stmt->bind_param('sss',$data['nombre'],$data['email'],$hashed_pass);
+            $stmt->bind_param('isssss',$data['nombre'],$data['email'],$hashed_pass);
             $stmt->execute();
 
             if($stmt->error){
@@ -159,9 +159,41 @@ class Usuarios{
         }
     }
 
+    //Update datos generales
+    public function updateUsuario($data,$email){
+        //Verifico si el usuario existe
+
+        $user = $this->getUserEmail($email);
+        if(empty($user)){
+            return ['message' => 'Usuario no encontrado o no existente'];
+        }
+        $query = 'UPDATE usuarios SET dni_usuario ?, nombre_usuario ?,nombreCompleto ?, num_telefono ? WHERE usuarios email = ?';
+
+        try {
+            $stmt = $this->con->prepare($query);
+            $stmt->bind_param('isss',$data['dni_usuario'],$data['nombre_usuario'],$data['nombreCompleto'],$data['num_telefono']);
+            $stmt->execute();
+
+            if($stmt->error){
+                throw new Exception('Error al actualizar datos');
+            }
+        } catch (\Throwable $th) {
+            return ['message' => $th->getMessage()];
+        }
+
+        return $stmt->affected_rows;
+
+    }
+
     //Eliminar usuario
     public function delete($email)
     {
+        //Verifico si el usuario existe
+        $user = $this->getUserEmail($email);
+        if(empty($user)){
+            return ['message' => 'Usuario no encontrado o no existente'];
+        }
+        
         $query = 'DELETE FROM usuarios WHERE email = ?';
 
         try {
